requirements.txt
google-generativeai
numpy
requests
